added transactionediting model and balance update method

+++ BugetApp/Models/Classes/Account.cs
@@ -7,7 +7,7 @@ namespace BugetApp.Models.Classes
         public int Id { get; set; }
 
         public string Name { get; set; }
-        public int Balance { get; set; }
+        public decimal Balance { get; set; }
 
         public int HouseholdId { get; set; }
         public virtual HouseHolds Household { get; set; }
